refactor: public api exceptions
