+++ src/data_factory_testing_framework/exceptions/__init__.py
@@ -1,24 +1,24 @@
-from data_factory_testing_framework.exceptions.activity_not_found_error import ActivityNotFoundError
-from data_factory_testing_framework.exceptions.activity_output_field_not_found_error import (
+from data_factory_testing_framework.exceptions._activity_not_found_error import ActivityNotFoundError
+from data_factory_testing_framework.exceptions._activity_output_field_not_found_error import (
     ActivityOutputFieldNotFoundError,
 )
-from data_factory_testing_framework.exceptions.data_factory_element_evaluation_error import (
+from data_factory_testing_framework.exceptions._data_factory_element_evaluation_error import (
     DataFactoryElementEvaluationError,
 )
-from data_factory_testing_framework.exceptions.function_call_invalid_arguments_count_error import (
+from data_factory_testing_framework.exceptions._function_call_invalid_arguments_count_error import (
     FunctionCallInvalidArgumentsCountError,
 )
-from data_factory_testing_framework.exceptions.parameter_not_found_error import ParameterNotFoundError
-from data_factory_testing_framework.exceptions.pipeline_activities_circular_dependency_error import (
+from data_factory_testing_framework.exceptions._parameter_not_found_error import ParameterNotFoundError
+from data_factory_testing_framework.exceptions._pipeline_activities_circular_dependency_error import (
     NoRemainingPipelineActivitiesMeetDependencyConditionsError,
 )
-from data_factory_testing_framework.exceptions.pipeline_not_found_error import PipelineNotFoundError
-from data_factory_testing_framework.exceptions.state_iteration_item_not_set_error import StateIterationItemNotSetError
-from data_factory_testing_framework.exceptions.unsupported_function_error import UnsupportedFunctionError
-from data_factory_testing_framework.exceptions.variable_being_evaluated_does_not_exist_error import (
+from data_factory_testing_framework.exceptions._pipeline_not_found_error import PipelineNotFoundError
+from data_factory_testing_framework.exceptions._state_iteration_item_not_set_error import StateIterationItemNotSetError
+from data_factory_testing_framework.exceptions._unsupported_function_error import UnsupportedFunctionError
+from data_factory_testing_framework.exceptions._variable_being_evaluated_does_not_exist_error import (
     VariableBeingEvaluatedDoesNotExistError,
 )
-from data_factory_testing_framework.exceptions.variable_not_found_error import VariableNotFoundError
+from data_factory_testing_framework.exceptions._variable_not_found_error import VariableNotFoundError
 
 __all__ = [
     "ActivityNotFoundError",
