+++ src/data_factory_testing_framework/exceptions/_activity_not_found_error.py
@@ -1,4 +1,4 @@
-from data_factory_testing_framework.exceptions.user_error import UserError
+from data_factory_testing_framework.exceptions._user_error import UserError
 
 
 class ActivityNotFoundError(UserError):
