+++ src/data_factory_testing_framework/exceptions/_function_call_invalid_arguments_count_error.py
@@ -1,4 +1,4 @@
-from data_factory_testing_framework.exceptions.user_error import UserError
+from data_factory_testing_framework.exceptions._user_error import UserError
 
 
 class FunctionCallInvalidArgumentsCountError(UserError):
