+++ src/data_factory_testing_framework/exceptions/_pipeline_activities_circular_dependency_error.py
@@ -1,4 +1,7 @@
-class NoRemainingPipelineActivitiesMeetDependencyConditionsError(Exception):
+from data_factory_testing_framework.exceptions._user_error import UserError
+
+
+class NoRemainingPipelineActivitiesMeetDependencyConditionsError(UserError):
     """NoRemainingPipelineActivitiesMeetDependencyConditionsError.
 
     This error is raised when there are still pending pipeline activities, but no conditions are met to execute these activities.
