+++ src/data_factory_testing_framework/exceptions/_state_iteration_item_not_set_error.py
@@ -1,4 +1,7 @@
-class StateIterationItemNotSetError(Exception):
+from data_factory_testing_framework.exceptions._user_error import UserError
+
+
+class StateIterationItemNotSetError(UserError):
     def __init__(self) -> None:
         """Error raised when an iteration item is not set."""
         super().__init__("Iteration item not set.")
