+++ src/data_factory_testing_framework/exceptions/_user_error.py
