+++ src/data_factory_testing_framework/exceptions/_variable_being_evaluated_does_not_exist_error.py
@@ -1,4 +1,7 @@
-class VariableBeingEvaluatedDoesNotExistError(Exception):
+from data_factory_testing_framework.exceptions._user_error import UserError
+
+
+class VariableBeingEvaluatedDoesNotExistError(UserError):
     def __init__(self, variable_name: str) -> None:
         """Error raised when a variable being evaluated does not exist."""
         super().__init__(f"Variable being evaluated does not exist: {variable_name}")
