+++ tests/functional/execute_child_pipeline/test_execute_pipeline_activity.py
@@ -1,6 +1,6 @@
 import pytest
 from data_factory_testing_framework import TestFramework, TestFrameworkType
-from data_factory_testing_framework.exceptions.pipeline_not_found_error import PipelineNotFoundError
+from data_factory_testing_framework.exceptions import PipelineNotFoundError
 from data_factory_testing_framework.state import RunParameter, RunParameterType
